router: Add method to server::Router to optimize search for all routes

+++ server/route.hpp
@@ -43,7 +43,7 @@ struct Route {
   Route_expr  expr;
   End_point   end_point;
   route::Keys keys;
-  unsigned    hits {0U};
+  size_t      hits {0U};
 }; //< struct Route
 
 inline bool operator < (const Route& lhs, const Route& rhs) noexcept {
